+++ src/main/java/com/github/ka4ok85/wca/command/ExportListCommand.java
@@ -86,7 +86,7 @@ public class ExportListCommand extends AbstractCommand<ExportListResponse, Expor
 			Node jobIdNode = (Node) xpath.evaluate("JOB_ID", resultNode, XPathConstants.NODE);
 			Node filePathNode = (Node) xpath.evaluate("FILE_PATH", resultNode, XPathConstants.NODE);
 
-			final int jobId = Integer.parseInt(jobIdNode.getTextContent());
+			final Long jobId = Long.parseLong(jobIdNode.getTextContent());
 			log.debug("Job ID {} is being excuted", jobId);
 			
 			final JobResponse jobResponse = waitUntilJobIsCompleted(jobId);
